REFACTOR: Replace number of return values with multiple return flag.
In TraceReturn events, that is. The number of return values can be ambiguous
in multiple, compound assignements, like `z = x, y = ...` (is it 1 or 2?),
and in any case we don't need to know a priori the number of return values,
which will be determined at runtime, but only whether to treat the return
value as a "multiple return".